Libtorrent y UnRAR: mejoras internas
