+++ plugin.video.alfa/platformcode/platformtools.py
@@ -1160,12 +1160,11 @@ def torrent_client_installed(show_tuple=False):
 def play_torrent(item, xlistitem, mediaurl):
     logger.info()
     import time
-    import threading
     import shutil
-    import os
     from core import filetools
     from core import httptools
     from lib import generictools
+    from servers import torrent
     
     # Opciones disponibles para Reproducir torrents
     torrent_options = list()
@@ -1190,6 +1189,10 @@ def play_torrent(item, xlistitem, mediaurl):
     # Si Libtorrent ha dado error de inicialización, no se pueden usar los clientes internos
     UNRAR = config.get_setting("unrar_path", server="torrent", default="")
     LIBTORRENT = config.get_setting("libtorrent_path", server="torrent", default='')
+    size_rar = 2
+    rar_files = []
+    if item.password:
+        size_rar = 3
     if seleccion < 2 and not LIBTORRENT:
         dialog_ok('Cliente Interno (LibTorrent):', 'Este cliente no está soportado en su dispositivo.',  \
                   'Error: [COLOR yellow]%s[/COLOR]' % config.get_setting("libtorrent_error", server="torrent", default=''), \
@@ -1200,12 +1203,12 @@ def play_torrent(item, xlistitem, mediaurl):
                 return
         else:
             return
-    elif seleccion > 1 and LIBTORRENT and UNRAR and ("torrenter" in torrent_options[seleccion][0] \
+    elif seleccion > 1 and LIBTORRENT and UNRAR and 'RAR-' in item.torrent_info and ("torrenter" in torrent_options[seleccion][0] \
                         or ("elementum" in torrent_options[seleccion][0] and xbmcaddon.Addon(id="plugin.video.%s" \
                         % torrent_options[seleccion][0].replace('Plugin externo: ', '')).getSetting('download_storage') == '1')):
         if dialog_yesno(torrent_options[seleccion][0], 'Este plugin externo no soporta extraer on-line archivos RAR',  \
                         '[COLOR yellow]¿Quiere que usemos esta vez el Cliente interno MCT?[/COLOR]', \
-                        'Esta operación ocupará en disco 2+ veces el tamaño del vídeo'):
+                        'Esta operación ocupará en disco [COLOR yellow][B]%s+[/B][/COLOR] veces el tamaño del vídeo' % size_rar):
                         seleccion = 1
         
     # Descarga de torrents a local
@@ -1255,7 +1258,7 @@ def play_torrent(item, xlistitem, mediaurl):
             if item.referer: referer = item.referer
             if item.post: post = item.post
             #Descargamos el .torrent
-            size, url, torrent, rar_files = generictools.get_torrent_size(item.url, referer, post, \
+            size, url, torrent_f, rar_files = generictools.get_torrent_size(item.url, referer, post, \
                         torrents_path=torrents_path, timeout=timeout, lookup=False, headers=headers)
             if url:
                 url_stat = True
@@ -1287,11 +1290,18 @@ def play_torrent(item, xlistitem, mediaurl):
         
         mediaurl = item.url
 
-    # Plugins externos
-    if seleccion > 0:
-        if seleccion == 1:
+    if seleccion >= 0:
+        
+        # Reproductor propio BT (libtorrent)
+        if seleccion == 0:
+            torrent.bt_client(mediaurl, xlistitem, rar_files, subtitle=item.subtitle, password=password, item=item)
+        
+        # Reproductor propio MCT (libtorrent)
+        elif seleccion == 1:
             from platformcode import mct
             mct.play(mediaurl, xlistitem, subtitle=item.subtitle, password=password, item=item)
+        
+        # Plugins externos
         else:
             mediaurl = urllib.quote_plus(item.url)
             #Llamada con más parámetros para completar el título
@@ -1307,12 +1317,11 @@ def play_torrent(item, xlistitem, mediaurl):
             # Si es un archivo RAR, monitorizamos el cliente Torrent hasta que haya descargado el archivo,
             # y después lo extraemos, incluso con RAR's anidados y con contraseña
             torr_client = torrent_options[seleccion][0].replace('Plugin externo: ', '')
-            if 'RAR-' in size and torr_client in ['quasar', 'elementum'] and \
-                                    config.get_setting("unrar_path", server="torrent", default=""):
-                rar_file, save_path_videos, folder_torr = wait_for_download(rar_files, torr_client)  # Esperamos mientras se descarga el RAR
+            if 'RAR-' in size and torr_client in ['quasar', 'elementum'] and UNRAR:
+                rar_file, save_path_videos, folder_torr = torrent.wait_for_download(rar_files, torr_client)  # Esperamos mientras se descarga el RAR
                 if rar_file and save_path_videos:                                       # Si se ha descargado el RAR...
                     dp = dialog_progress_bg('Alfa %s' % torr_client)
-                    video_file, rar, video_path, erase_file_path = extract_files(rar_file, \
+                    video_file, rar, video_path, erase_file_path = torrent.extract_files(rar_file, \
                                     save_path_videos, password, dp, item, torr_client)  # ... extraemos el vídeo del RAR
                     dp.close()
                     
@@ -1328,700 +1337,28 @@ def play_torrent(item, xlistitem, mediaurl):
                         playlist.add(video_play, xlistitem)
                         xbmc_player.play(playlist)
 
-        #Seleccionamos que clientes torrent soportamos para el marcado de vídeos vistos: asumimos que todos funcionan
-        #if "quasar" in torrent_options[seleccion][1] or "elementum" in torrent_options[seleccion][1]:   
-
-        time_limit = time.time() + 150                                      #Marcamos el timepo máx. de buffering
-        while not is_playing() and time.time() < time_limit:                #Esperamos mientra buffera    
-            time.sleep(5)                                                   #Repetimos cada intervalo
-            #logger.debug(str(time_limit))
-        if item.subtitle != '':
-            time.sleep(5)
-            xbmc_player.setSubtitles(item.subtitle)
-            #subt = xbmcgui.ListItem(path=item.url, thumbnailImage=item.thumbnail)
-            #subt.setSubtitles([item.subtitle])
-
-        if item.strm_path and is_playing():                                 #Sólo si es de Videoteca
-            from platformcode import xbmc_videolibrary
-            xbmc_videolibrary.mark_auto_as_watched(item)                    #Marcamos como visto al terminar
-            #logger.debug("Llamado el marcado")
-            
-        # Si se ha extraido un RAR, se pregunta para borrar los archivos después de reproducir el vídeo
-        while is_playing() and rar and not xbmc.abortRequested:
-            time.sleep(3)                                                   #Repetimos cada intervalo
-        if rar and not xbmc.abortRequested:
-            if dialog_yesno('Alfa %s' % torr_client, '¿Borrar las descargas del RAR y Vídeo?'):
-                log("##### erase_file_path: %s" % erase_file_path)
-                try:
-                    torr_data, deamon_url, index = get_tclient_data(folder_torr, torr_client)
-                    if torr_data and deamon_url:
-                        data = httptools.downloadpage('%sdelete/%s' % (deamon_url, index), timeout=5, alfa_s=True).data
-                    shutil.rmtree(erase_file_path, ignore_errors=True)
-                except:
-                    logger.error(traceback.format_exc(1))
-
-    # Reproductor propio (libtorrent)
-    if seleccion == 0:
-        import time
-        played = False
-        debug = (config.get_setting("debug") == True)
-
-        # Importamos el cliente
-        from btserver import Client
-
-        save_path_videos = filetools.join(config.get_setting("bt_download_path", server="torrent", 
-                   default=config.get_setting("downloadpath")), 'BT-torrents')
-        if not save_path_videos:
-            save_path_videos = filetools.join(config.get_data_path(), 'BT-torrents')
-            
-        BACKGROUND = config.get_setting("mct_background_download", server="torrent", default=True)
-        RAR = config.get_setting("mct_rar_unpack", server="torrent", default=True)
-        try:
-            BUFFER = int(config.get_setting("bt_buffer", server="torrent", default="50"))
-        except:
-            BUFFER = 50
-        torr_client = 'BT'
-        rar_file = ''
-        rar_names = []
-        rar = False
-        video_names = []
-        video_file = ''
-        video_path = ''
-        videourl = ''
-        msg_header = 'Alfa %s Cliente Torrent' % torr_client
-        extensions_list = ['.aaf', '.3gp', '.asf', '.avi', '.flv', '.mpeg',
-                           '.m1v', '.m2v', '.m4v', '.mkv', '.mov', '.mpg',
-                           '.mpe', '.mp4', '.ogg', '.rar', '.wmv', '.zip']
-        for entry in rar_files:
-            for file, path in entry.items():
-                if file == 'path' and '.rar' in str(path):
-                    for file_r in path:
-                        rar_names += [file_r]
-                        rar = True
-                elif file == 'path' and not '.rar' in str(path):
-                    for file_r in path:
-                        if os.path.splitext(file_r)[1] in extensions_list:
-                            video_names += [file_r]
-                elif file == '__name':
-                    video_path = path
-        if rar: rar_file = '%s/%s' % (video_path, rar_names[0])
-        erase_file_path = filetools.join(save_path_videos, video_path)
-        video_path = erase_file_path
-        if video_names: video_file = video_names[0]
-        
-        if rar and RAR and not UNRAR:
-            if not dialog_yesno(msg_header, 'Se ha detectado un archivo .RAR en la descarga', \
-                        'No tiene instalado el extractor UnRAR', '¿Desea descargarlo en cualquier caso?'):
-                return
-
-        # Iniciamos el cliente:
-        c = Client(url=mediaurl, is_playing_fnc=xbmc_player.isPlaying, wait_time=None, timeout=10,
-                   temp_path=save_path_videos, print_status=debug, auto_delete=False)
-        activo = True
-        finalizado = False
-
-        # Mostramos el progreso
-        if rar and RAR and BACKGROUND:                                                  # Si se descarga un RAR...
-            progreso = dialog_progress_bg(msg_header)
-            dialog_notification("Descarga de RAR en curso", "Puedes realizar otras tareas en Kodi mientrastanto. " + \
-                    "Te informaremos...", time=10000)
-        else:
-            progreso = dialog_progress('Alfa %s Cliente Torrent' % torr_client, '')
-
-        # Mientras el progreso no sea cancelado ni el cliente cerrado
-        while not c.closed and not xbmc.abortRequested:
-            try:
-                # Obtenemos el estado del torrent
-                s = c.status
-                if debug:
-                    # Montamos las tres lineas con la info del torrent
-                    txt = '%.2f%% de %.1fMB %s | %.1f kB/s' % \
-                          (s.progress_file, s.file_size, s.str_state, s._download_rate)
-                    txt2 = 'S: %d(%d) P: %d(%d) | DHT:%s (%d) | Trakers: %d | Pi: %d(%d)' % \
-                           (s.num_seeds, s.num_complete, s.num_peers, s.num_incomplete, s.dht_state, s.dht_nodes,
-                            s.trackers, s.pieces_sum, s.pieces_len)
-                    txt3 = 'Origen Peers TRK: %d DHT: %d PEX: %d LSD %d ' % \
-                           (s.trk_peers, s.dht_peers, s.pex_peers, s.lsd_peers)
-                else:
-                    txt = '%.2f%% de %.1fMB %s | %.1f kB/s' % \
-                          (s.progress_file, s.file_size, s.str_state, s._download_rate)
-                    txt2 = 'S: %d(%d) P: %d(%d)' % (s.num_seeds, s.num_complete, s.num_peers, s.num_incomplete)
-                    try:
-                        txt3 = config.get_localized_string(70197) % (int(s.timeout))
-                    except:
-                        txt3 = ''
-
-                if rar and RAR and BACKGROUND:
-                    progreso.update(s.buffer, txt, txt2)
-                else:
-                    progreso.update(s.buffer, txt, txt2, txt3)
-                time.sleep(1)
-
-                if not (rar and RAR and BACKGROUND) and progreso.iscanceled():
-                    progreso.close()
-                    if s.buffer == 100 or 'Finalizado' in s.str_state or 'Seeding' in s.str_state:
-                        if not rar and dialog_yesno(msg_header, config.get_localized_string(70198)):
-                            played = False
-                            progreso = dialog_progress(msg_header, '')
-                            progreso.update(s.buffer, txt, txt2, txt3)
-                        else:
-                            progreso = dialog_progress(msg_header, '')
-                            break
-
-                    else:
-                        if dialog_yesno(msg_header, config.get_localized_string(70199)):
-                            progreso = dialog_progress(msg_header, '')
-                            break
-
-                        else:
-                            progreso = dialog_progress(msg_header, '')
-                            progreso.update(s.buffer, txt, txt2, txt3)
-
-                # Si el buffer se ha llenado y la reproduccion no ha sido iniciada, se inicia
-                if ((s.pieces_sum >= BUFFER  or 'Finalizado' in s.str_state or 'Seeding' in s.str_state) and not rar) or \
-                            (s.pieces_sum == s.pieces_len and s.pieces_len > 0 and ('Finalizado' in s.str_state or 'Seeding' \
-                            in s.str_state) and rar) and not played:
-                    
-                    if rar and RAR and UNRAR:
-                        c.stop()
-                        activo = False
-                        finalizado = True
-                        video_file, rar_res, video_path, erase_file_path = extract_files(rar_file, \
-                                        save_path_videos, password, progreso, item, torr_client)  # ... extraemos el vídeo del RAR
-                        if rar_res and not xbmc.abortRequested:
-                            time.sleep(1)
-                        else:
-                            break
-                    elif (rar and not UNRAR) or (rar and not RAR):
-                        break
-                    
-                    # Cerramos el progreso
-                    progreso.close()
-
-                    # Reproducimos el vídeo extraido, si no hay nada en reproducción
-                    while xbmc_player.isPlaying() and not xbmc.abortRequested:
-                        time.sleep(3)      
-                    
-                    # Obtenemos el playlist del torrent
-                    #videourl = c.get_play_list()
-                    videourl = filetools.join(video_path, video_file)
-
-                    # Iniciamos el reproductor
-                    playlist = xbmc.PlayList(xbmc.PLAYLIST_VIDEO)
-                    playlist.clear()
-                    playlist.add(videourl, xlistitem)
-                    # xbmc_player = xbmc_player
-                    log("##### videourl: %s" % videourl)
-                    xbmc_player.play(playlist)
-
-                    # Marcamos como reproducido para que no se vuelva a iniciar
-                    played = True
-                    
-                    if item.subtitle != '':
-                        time.sleep(5)
-                        xbmc_player.setSubtitles(item.subtitle)
-
-                    # si es un archivo de la videoteca enviar a marcar como visto
-                    if item.strm_path:
-                        from platformcode import xbmc_videolibrary
-                        xbmc_videolibrary.mark_auto_as_watched(item)
-
-                    # Y esperamos a que el reproductor se cierre
-                    while xbmc_player.isPlaying() and not xbmc.abortRequested:
-                        time.sleep(3)
-                    
-                    # Miramos si se ha completado la descarga para borrar o no los archivos
-                    if activo:
-                        s = c.status
-                    if s.pieces_sum == s.pieces_len:
-                        finalizado = True
-
-                    # Cuando este cerrado,  Volvemos a mostrar el dialogo
-                    if not rar:
-                        progreso = dialog_progress(msg_header, '')
-                        progreso.update(s.buffer, txt, txt2, txt3)
-                        
-                    break
-
-            except:
-                import traceback
-                logger.error(traceback.format_exc())
-                break
-
-        if rar:
-            progreso.update(100, config.get_localized_string(70200), " ")
-        else:
-            progreso.update(100, config.get_localized_string(70200), " ", " ")
-
-        # Detenemos el cliente
-        if activo and not c.closed:
-            c.stop()
-            activo = False
-
-        # Cerramos el progreso
-        progreso.close()
-        
-        # Y borramos los archivos de descarga restantes
-        time.sleep(1)
-        if filetools.exists(erase_file_path):
-            if finalizado and not dialog_yesno(msg_header, '¿Borrar los archivos descargados?'):
-                return
-            log("##### erase_file_path: %s" % erase_file_path)
-            for x in range(10):
-                shutil.rmtree(erase_file_path, ignore_errors=True)
-                time.sleep(5)
-                if not filetools.exists(erase_file_path):
-                    break
-
-
-def wait_for_download(rar_files, torr_client):
-    logger.info()
-    import time
-    import traceback
-    from subprocess import Popen, PIPE, STDOUT
-    from core import filetools, scrapertools
-    
-    # Analizamos los archivos dentro del .torrent
-    rar = False
-    rar_names = []
-    rar_names_abs = []
-    folder = ''
-    for entry in rar_files:
-        for file, path in entry.items():
-            if file == 'path' and '.rar' in str(path):
-                for file_r in path:
-                    rar_names += [file_r]
-                    rar = True
-            elif file == '__name':
-                folder = path
-    
-    if not folder:                                                              # Si no se detecta el folder...
-        return ('', '', '')                                                     # ... no podemos hacer nada
-        
-    if not rar_names:
-        return ('', '', folder)
-    rar_file = '%s/%s' % (folder, rar_names[0])
-    log("##### rar_file: %s" % rar_file)
-    if len(rar_names) > 1:
-        log("##### rar_names: %s" % str(rar_names))
-        
-    # Localizamos el path de descarga del .torrent
-    save_path_videos = ''
-    __settings__ = xbmcaddon.Addon(id="plugin.video.%s" % torr_client)          # Apunta settings del cliente torrent
-    if torr_client == 'torrenter':
-        save_path_videos = str(xbmc.translatePath(__settings__.getSetting('storage')))
-        if not save_path_videos:
-            save_path_videos = str(filetools.join(xbmc.translatePath("special://home/"), \
-                                   "cache", "xbmcup", "plugin.video.torrenter", "Torrenter"))
-    else:
-        save_path_videos = str(xbmc.translatePath(__settings__.getSetting('download_path')))
-        if __settings__.getSetting('download_storage') == '1':                  # Descarga en memoria?
-            return ('', '', folder)                                             # volvemos
-    if not save_path_videos:                                                    # No hay path de descarga?
-        return ('', '', folder)                                                 # Volvemos
-    log("##### save_path_videos: %s" % save_path_videos)
-
-    # Esperamos mientras el .torrent se descarga.  Verificamos si el .RAR está descargado al completo
-    dialog_notification("Automatizando la extracción", "Acepta descargar el archivo RAR y te iremos guiando...", time=10000)
-    
-    # Plan A: usar el monitor del cliente torrent para ver el status de la descarga
-    loop = 3600                                                                 # Loop de 10 horas hasta crear archivo
-    wait_time = 10
-    time.sleep(wait_time)
-    fast = False
-    for x in range(loop):
-        if xbmc.abortRequested:
-            return ('', '', folder)
-        torr_data, deamon_url, index = get_tclient_data(folder, torr_client)
-        if not torr_data or not deamon_url:
-            break
-        if torr_client in ['quasar'] and not torr_data['label'].startswith('0.00%') and not fast:
-            dialog_notification("Descarga en curso", "Puedes realizar otras tareas en Kodi mientrastanto. " + \
-                    "Te informaremos...", time=10000)
-            fast = True
-        if not torr_data['label'].startswith('100.00%'):
-            log("##### Descargado: %s, ID: %s" % (scrapertools.find_single_match(torr_data['label'], '(^.*?\%)'), index))
-            time.sleep(wait_time)
-            continue
-        
-        log("##### Torrent FINALIZADO: %s" % str(folder))
-        return (rar_file, save_path_videos, folder)
-    
-    # Plan B: monitorizar con UnRAR si los archivos se han desacargado por completo
-    unrar_path = config.get_setting("unrar_path", server="torrent", default="")
-    if not unrar_path:                                                          # Si Unrar no está instalado...
-        return ('', '', folder)                                                 # ... no podemos hacer nada
-        
-    cmd = []
-    for rar_name in rar_names:                                                  # Preparamos por si es un archivo multiparte
-        cmd.append(['%s' % unrar_path, 'l', '%s' % filetools.join(save_path_videos, folder, rar_name)])
-    
-    creationflags = ''
-    if xbmc.getCondVisibility("system.platform.Windows"):
-        creationflags = 0x08000000
-    loop = 30                                                                   # Loop inicial de 5 minutos hasta crear archivo
-    wait_time = 10
-    loop_change = 0
-    loop_error = 6
-    part_name = ''
-    y = 0
-    returncode = ''
-    fast = False
-    while rar and not xbmc.abortRequested:
-        for x in range(loop):                                                   # Loop corto (5 min.) o largo (10 h.)
-            if xbmc.abortRequested:
-                return ('', '', folder)
-            if not rar or loop_change > 0:
-                loop = loop_change                                              # Paso de loop corto a largo
-                loop_change = 0
-                break
-            try:
-                responses = []
-                for z, command in enumerate(cmd):                               # Se prueba por cada parte
-                    data_rar = Popen(command, bufsize=0, stdout=PIPE, stdin=PIPE, \
-                                     stderr=STDOUT, creationflags=creationflags)
-                    out_, error_ = data_rar.communicate()
-                    responses.append([z, str(data_rar.returncode), out_, error_])   # Se guarda la respuesta de cada parte
-            except:
-                logger.error(traceback.format_exc(1))                           # Error de incompatibilidad de UnRAR
-                rar = False
-                break
-            else:
-                dl_files = 0
-                for z, returncode, out__, error__ in responses:                 # Analizamos las respuestas
-                    if returncode == '0':                                       # Ya se ha descargado... parte ...
-                        dl_files += 1
-                        part_name = scrapertools.find_single_match(str(out__), '(\.part\d+.rar)')
-                        log("##### Torrent descargando: %s, %s" % (part_name, str(returncode)))
-                        if dl_files == len(cmd):                                # ... o todo
-                            fast = True
-                            rar = False
-                            break                                                   # ... o sólo una parte
-                    elif returncode == '10':                                    # archivo no existe
-                        if loop != 30:                                          # Si el archivo es borrado durante el proceso ...
-                            rar = False
-                        break                                                   #... abortamos
-                    elif returncode == '6':                                     # En proceso de descarga
-                        y += 1
-                        #if loop == 30 and y == len(responses):                  # Si es la primera vez en proceso ...
-                        if loop == 30 and y == 1:                               # Si es la primera vez en proceso ...
-                            if torr_client in ['quasar']:
-                                dialog_notification("Descarga en curso", "Puedes realizar otras tareas en Kodi mientrastanto. " + \
-                                        "Te informaremos...", time=10000)
-                            loop_change = 3600                                  # ... pasamos a un loop de 10 horas
-                        elif loop <= 6:                                         # Recuerado el error desconocido
-                            loop_change = 3600                                  # ... pasamos a un loop de 10 horas
-                            loop_error = 6                                      # Restauramos loop_error por si acaso
-                        break
-                    elif returncode == '1':                                     # Ha alcanzado el fin de archivo ??? pasamos
-                        part_name = scrapertools.find_single_match(str(out__), '(\.part\d+.rar)')
-                        log("##### Torrent descargando: %s, %s" % (part_name, str(returncode)))
-                    else:                                                       # No entendemos el error
-                        loop_change = loop_error                                # ... pasamos a un loop de 1 minutos para reintentar
-                        loop_error += -1
-                        break                                                   #... abortamos
+        if seleccion > 1:
+            #Seleccionamos que clientes torrent soportamos para el marcado de vídeos vistos: asumimos que todos funcionan
+            torrent.mark_auto_as_watched(item)
                 
-                if str(returncode) in ['0', '6', '10']:
-                    log("##### Torrent descargando: %s" % str(returncode))
-                else:
-                    log("##### Torrent descargando: %s, %s" % (str(out__), str(returncode)))
-                if not rar or fast:
-                    fast = False
-                    break
-                time.sleep(wait_time)                                           # Esperamos un poco y volvemos a empezar
-        else:
-            rar = False
-            break
-
-    if str(returncode) == '0':
-        log("##### Torrent FINALIZADO: %s" % str(returncode))
-    else:
-        rar_file = ''
-        logger.error('##### Torrent NO DESCARGADO: %s, %s' % (str(out__), str(returncode)))
-    
-    return (rar_file, save_path_videos, folder)
-    
-    
-def get_tclient_data(folder, torr_client):
-    import json
-    import traceback
-    from core import httptools, scrapertools
-    
-    # Monitoriza el estado de descarga del torrent en Quasar y Elementum
-    local_host = {"quasar": "http://localhost:65251/torrents/", "elementum": "http://localhost:65220/torrents/"}
-    torr = ''
-    torr_id = ''
-    x = 0
-    y = ''
-    
-    try:
-        data = httptools.downloadpage(local_host[torr_client], timeout=5, alfa_s=True).data
-        if not data:
-            return '', local_host[torr_client], 0
-
-        data = json.loads(data)
-        data = data['items']
-        for x, torr in enumerate(data):
-            if not folder in torr['label']:
-                continue
-            if "elementum" in torr_client:
-                torr_id = scrapertools.find_single_match(str(torr), 'torrents\/pause\/(.*?)\)')
-    except:
-        log(traceback.format_exc(1))
-        return '', local_host[torr_client], 0
-    
-    if torr_id:
-        y = torr_id
-    else:
-        y = x
-    return torr, local_host[torr_client], y
-
-
-def extract_files(rar_file, save_path_videos, password, dp, item=None, torr_client=None):
-    logger.info()
-    import rarfile
-    import sys
-    reload(sys)
-    sys.setdefaultencoding('latin1')
-    import traceback
-    from core import filetools
-
-    # Verificamos si hay path para UnRAR
-    rarfile.UNRAR_TOOL = config.get_setting("unrar_path", server="torrent", default="")
-    if not rarfile.UNRAR_TOOL:
-        return rar_file, False, '', ''
-    log("##### unrar_path: %s" % rarfile.UNRAR_TOOL)
-    if xbmc.getCondVisibility("system.platform.Android"):
-        rarfile.UNRAR_TOOL = xbmc.executebuiltin("StartAndroidActivity(com.rarlab.rar)")
-    rarfile.DEFAULT_CHARSET = 'latin1'
-    
-    # Preparamos un path alternativo más corto para no sobrepasar la longitud máxima
-    video_path = ''
-    if item:
-        if item.contentType == 'movie':
-            video_path = '%s-%s' % (item.contentTitle, item.infoLabels['tmdb_id'])
-        else:
-            video_path = '%s-%sx%s-%s' % (item.contentSerieName, item.contentSeason, \
-                            item.contentEpisodeNumber, item.infoLabels['tmdb_id'])
-    
-    # Renombramos el path dejado en la descarga a uno más corto
-    rename_status = False
-    org_rar_file = rar_file
-    org_save_path_videos = save_path_videos
-    if video_path and '/' in rar_file:
-        log("##### rar_file: %s" % rar_file)
-        rename_status, rar_file = rename_rar_dir(org_rar_file, org_save_path_videos, video_path, torr_client)
-
-    # Calculamos el path para del RAR
-    if "/" in rar_file:
-        folders = rar_file.split("/")
-        erase_file_path = filetools.join(save_path_videos, folders[0])
-        file_path = save_path_videos
-        for f in folders:
-            file_path = filetools.join(file_path, f)
-    else:
-        file_path = save_path_videos
-        erase_file_path = save_path_videos
-
-    # Calculamos el path para la extracción
-    if "/" in rar_file:
-        folders = rar_file.split("/")
-        for f in folders:
-            if not '.rar' in f:
-                save_path_videos = filetools.join(save_path_videos, f)
-    save_path_videos = filetools.join(save_path_videos, 'Extracted')
-    if not filetools.exists(save_path_videos): filetools.mkdir(save_path_videos)
-    log("##### save_path_videos: %s" % save_path_videos)
-
-    # Permite hasta 5 pasadas de extracción de .RARs anidados
-    dialog_notification("Empezando extracción...", rar_file, time=5000)
-    for x in range(5):
-        try:
-            archive = rarfile.RarFile(file_path.decode("utf8"))
-        except:
-            log("##### ERROR en Archivo rar: %s" % rar_file)
-            log("##### ERROR en Carpeta del rar: %s" % file_path)
-            log(traceback.format_exc(1))
-            dialog_notification("Error al abrir el RAR", "Comprueba el log para más detalles")
-            return rar_file, False, '', ''
-
-        # Analizamos si es necesaria una contraseña, que debería estar en item.password
-        if archive.needs_password():
-            if not password:
-                pass_path = filetools.split(file_path)[0]
-                password = last_password_search(pass_path)
-            if not password :
-                password = dialog_input(heading="Introduce la contraseña (Mira en %s)" % pass_path)
-                if not password:
-                    return rar_file, False, '', ''
-            archive.setpassword(password)
-
-        # Miramos el contenido del RAR a extraer
-        files = archive.infolist()
-        info = []
-        for idx, i in enumerate(files):
-            if i.file_size == 0:
-                files.pop(idx)
-                continue
-            filename = i.filename
-            if "/" in filename:
-                filename = filename.rsplit("/", 1)[1]
-
-            info.append("%s - %.2f MB" % (filename, i.file_size / 1048576.0))
-        if info:
-            info.append("Extraer todo sin reproducir")
-        else:
-            dialog_notification("El RAR está vacío", "O no contiene archivos válidos")
-            return rar_file, False, '', erase_file_path
-
-        # Seleccionamos extraer TODOS los archivos del RAR
-        #selection = xbmcgui.Dialog().select("Selecciona el fichero a extraer y reproducir", info)
-        selection = len(info) - 1
-        if selection < 0:
-            return rar_file, False, '', erase_file_path
-        else:
-            try:
-                log("##### RAR Extract INI #####")
-                if selection == len(info) - 1:
-                    log("##### rar_file 1: %s" % file_path)
-                    log("##### save_path_videos 1: %s" % save_path_videos)
-                    dp.update(99, "Extrayendo archivos...", "Espera unos minutos....")
-                    archive.extractall(save_path_videos)
-                else:
-                    log("##### rar_file 2: %s" % file_path)
-                    log("##### save_path_videos 2: %s" % save_path_videos)
-                    dp.update(99, "Espera unos minutos....", "Extrayendo archivo... %s" % info[selection])
-                    archive.extract(files[selection], save_path_videos)
-                log("##### RAR Extract END #####")
-            except (rarfile.RarWrongPassword, rarfile.RarCRCError):
-                dialog_notification("Error al extraer", "Contraseña incorrecta")
-                log(traceback.format_exc(1))
-                return rar_file, False, '', erase_file_path
-            except rarfile.BadRarFile:
-                dialog_notification("Error al extraer", "Archivo rar con errores")
-                log(traceback.format_exc(1))
-                return rar_file, False, '', erase_file_path
-            except:
-                dialog_notification("Error al extraer", "Comprueba el log para más detalles")
-                log(traceback.format_exc(1))
-                return rar_file, False, '', erase_file_path
-
-            extensions_list = ['.aaf', '.3gp', '.asf', '.avi', '.flv', '.mpeg',
-                               '.m1v', '.m2v', '.m4v', '.mkv', '.mov', '.mpg',
-                               '.mpe', '.mp4', '.ogg', '.wmv']
-            
-            # Localizamos el path donde se ha dejado la extracción
-            folder = True
-            file_result = filetools.listdir(save_path_videos)
-            while folder:
-                for file_r in file_result:
-                    if filetools.isdir(filetools.join(save_path_videos, file_r)):
-                        file_result_alt = filetools.listdir(filetools.join(save_path_videos, file_r))
-                        if file_result_alt:
-                            file_result = file_result_alt
-                            save_path_videos = filetools.join(save_path_videos, file_r)
-                        else:
-                            folder = False
-                        break
-                else:
-                    folder = False
-
-            # Si hay RARs anidados, ajustamos los paths para la siguiente pasada
-            if '.rar' in str(file_result):
-                for file_r in file_result:
-                    if '.rar' in file_r:
-                        rar_file = file_r
-                        file_path = str(filetools.join(save_path_videos, rar_file))
-                        save_path_videos = filetools.join(save_path_videos, 'Extracted')
-                        if not filetools.exists(save_path_videos): filetools.mkdir(save_path_videos)
-                        dialog_notification("Siguiente extracción...", rar_file, time=5000)
-            
-            # Si ya se ha extraido todo, preparamos el retorno            
-            else:
-                video_list = []
-                for file_r in file_result:
-                    if os.path.splitext(file_r)[1] in extensions_list:
-                        video_list += [file_r]
-                if len(video_list) == 0:
-                    dialog_notification("El rar está vacío", "O no contiene archivos válidos")
-                    return rar_file, False, '', erase_file_path
-                else:
-                    log("##### Archivo extraído: %s" % video_list[0])
-                    dialog_notification("Archivo extraído...", video_list[0], time=10000)
-                    return str(video_list[0]), True, save_path_videos, erase_file_path
-
-
-def rename_rar_dir(rar_file, save_path_videos, video_path, torr_client):
-    logger.info()
-    import traceback
-    from core import filetools, httptools
-
-    rename_status = False
-    folders = rar_file.split("/")
-    if filetools.exists(filetools.join(save_path_videos, folders[0])):
-        src = filetools.join(save_path_videos, folders[0]).decode("utf8")
-        dst = filetools.join(save_path_videos, video_path).decode("utf8")
-        
-        # Se para la actividad para que libere los archivos descargados
-        if torr_client in ['quasar', 'elementum']:
-            torr_data, deamon_url, index = get_tclient_data(folders[0], torr_client)
-            log("##### Client URL: %s" % '%spause/%s' % (deamon_url, index))
-            if torr_data and deamon_url:
-                data = httptools.downloadpage('%spause/%s' % (deamon_url, index), timeout=5, alfa_s=True).data
-            
-        for x in range(60):
-            if xbmc.abortRequested:
-                return rename_status, rar_file
-            xbmc.sleep(5000)
-            try:
-                if filetools.exists(src):
-                    filetools.rename(src, dst)
-                else:
-                    log("##### Folder DELETED: SRC: %s" % src)
-                    break
-            except:
-                log("##### Rename ERROR: SRC: %s" % src)
-                log(traceback.format_exc(1))
-            else:
-                if filetools.exists(dst):
-                    log("##### Renamed: SRC: %s" % src)
-                    log("##### TO: DST: %s" % dst)
-                    rar_file = video_path + '/' + folders[1]
-                    rename_status = True
-                    break
-                    
-    return rename_status, rar_file
-
-
-def last_password_search(pass_path):
-    logger.info()
-    import traceback
-    import re
-    from core import filetools, scrapertools, httptools
-    
-    # Busca en el Path de extracción si hay algún archivo que contenga la URL donde pueda estar la CONTRASEÑA
-    password = ''
-    patron_url = '(http.*\:\/\/(?:www.)?\w+\.\w+\/.*?[\n|\r|$])'
-    patron_pass = '<input\s*type="text"\s*id="txt_password"\s*name="[^"]+"\s*onClick="[^"]+"\s*value="([^"]+)"'
-    
-    try:
-        pass_path_list = filetools.listdir(pass_path)
-        for file in pass_path_list:
-            if 'contrase' in file.lower() and '.rar' not in file:
-                file_pass = filetools.read(filetools.join(pass_path, file))
-                url = scrapertools.find_single_match(file_pass, patron_url)
-                if url:
-                    data = re.sub(r"\n|\r|\t|\s{2}|(<!--.*?-->)", "", httptools.downloadpage(url).data)
-                    password = scrapertools.find_single_match(data, patron_pass)
-            if password:
-                break
-    except:
-        log(traceback.format_exc(1))
-    
-    log("##### Contraseña extraída: %s" % password)
-    return password
+            # Si se ha extraido un RAR, se pregunta para borrar los archivos después de reproducir el vídeo (plugins externos)
+            while is_playing() and rar and not xbmc.abortRequested:
+                time.sleep(3)                                               #Repetimos cada intervalo
+            if rar and not xbmc.abortRequested:
+                if dialog_yesno('Alfa %s' % torr_client, '¿Borrar las descargas del RAR y Vídeo?'):
+                    log("##### erase_file_path: %s" % erase_file_path)
+                    try:
+                        torr_data, deamon_url, index = torrent.get_tclient_data(folder_torr, torr_client)
+                        if torr_data and deamon_url:
+                            data = httptools.downloadpage('%sdelete/%s' % (deamon_url, index), timeout=5, alfa_s=True).data
+                        time.sleep(1)
+                        if filetools.isdir(erase_file_path):
+                            shutil.rmtree(erase_file_path, ignore_errors=True)
+                        elif filetools.exists(erase_file_path) and filetools.isfile(erase_file_path):
+                            filetools.remove(erase_file_path)
+                    except:
+                        logger.error(traceback.format_exc(1))
 
 
 def log(texto):
     xbmc.log(texto, xbmc.LOGNOTICE)
-
